Add files via upload
I wrote this ETL script when I received a list Twitter, Facebook, and Instagram which needed to be uploaded to a server of one of the applications we used to monitor social media activity. All of the authors came embedded in messy URLs. I extracted each author name from every URL.